Detect author from "references" if not already provided
Medium seems to be leaving the "displayAuthor" field blank and
keeping the info in a separate "references" field instead. Now,
we'll start pulling the name from there (and as a bonus, also
return a "list of authors" with username, name and userId, in
case someone else (ie. me) wants to do additional processing with
on the data