Automatically generate slug if not provided
This is an issue with Medium drafts, where a slug is not set yet.
mediumexporter tries to save it using an empty string (""), which
results in a file named `.md` or the `images` folder being created
in the main content directory rather than in a subfolder.

By automatically setting the slug, mediumexporter can ensure that
each story gets saved in its own  directory